# docs/api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/register_peer": {
      "post": {
        "summary": "Register Peer",
        "operationId": "register_peer_register_peer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeerRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/peers": {
      "get": {
        "summary": "Get Peers",
        "operationId": "get_peers_peers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/transaction": {
      "post": {
        "summary": "Add Transaction",
        "operationId": "add_transaction_transaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receive_transaction": {
      "post": {
        "summary": "Receive Transaction",
        "operationId": "receive_transaction_receive_transaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionReceiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/block": {
      "post": {
        "summary": "Receive Block",
        "operationId": "receive_block_block_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockReceiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chain": {
      "get": {
        "summary": "Get Chain",
        "operationId": "get_chain_chain_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChainBlockResponse"
                  },
                  "type": "array",
                  "title": "Response Get Chain Chain Get"
                }
              }
            }
          }
        }
      }
    },
    "/mine": {
      "post": {
        "summary": "Mine",
        "operationId": "mine_mine_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MineResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlockReceiveRequest": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionReceiveRequest"
            },
            "type": "array",
            "title": "Transactions"
          },
          "previous_hash": {
            "type": "string",
            "title": "Previous Hash"
          },
          "nonce": {
            "type": "integer",
            "title": "Nonce"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "zk_proof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zk Proof"
          }
        },
        "type": "object",
        "required": [
          "index",
          "transactions",
          "previous_hash",
          "nonce",
          "hash"
        ],
        "title": "BlockReceiveRequest"
      },
      "BlockResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "BlockResponse"
      },
      "ChainBlockResponse": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "tx": {
            "type": "integer",
            "title": "Tx"
          }
        },
        "type": "object",
        "required": [
          "index",
          "hash",
          "tx"
        ],
        "title": "ChainBlockResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MineResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hash"
          },
          "zk_proof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zk Proof"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MineResponse"
      },
      "PeerListResponse": {
        "properties": {
          "peers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Peers"
          }
        },
        "type": "object",
        "required": [
          "peers"
        ],
        "title": "PeerListResponse"
      },
      "PeerRegistrationRequest": {
        "properties": {
          "address": {
            "type": "string",
            "minLength": 1,
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "PeerRegistrationRequest"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusResponse"
      },
      "TransactionCreateRequest": {
        "properties": {
          "receiver": {
            "type": "string",
            "minLength": 1,
            "title": "Receiver"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "receiver",
          "amount"
        ],
        "title": "TransactionCreateRequest"
      },
      "TransactionCreatedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          }
        },
        "type": "object",
        "required": [
          "status",
          "sender"
        ],
        "title": "TransactionCreatedResponse"
      },
      "TransactionReceiveRequest": {
        "properties": {
          "sender": {
            "type": "string",
            "minLength": 1,
            "title": "Sender"
          },
          "receiver": {
            "type": "string",
            "minLength": 1,
            "title": "Receiver"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "sender",
          "receiver",
          "amount"
        ],
        "title": "TransactionReceiveRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}